Refactor package to io.opentelemetry.exporter.otlp.http.metrics for consistency
