+++ exporters/otlp-http/metrics/src/main/java/io/opentelemetry/exporter/otlp/http/metrics/package-info.java
@@ -5,6 +5,6 @@
 
 /** OpenTelemetry exporter which sends metric data to OpenTelemetry collector via OTLP HTTP. */
 @ParametersAreNonnullByDefault
-package io.opentelemetry.exporter.otlp.http.metric;
+package io.opentelemetry.exporter.otlp.http.metrics;
 
 import javax.annotation.ParametersAreNonnullByDefault;
